test: test nan with always_nearest
